fix(overview): remove the refreshing after posting a post
With the new stream implementation of getNearPosts, the post will directly appear in the feed after being posted without the need of a refresh.
This means that the refresh button will not appear.
The scroll down to refresh is also removed because we want to check that the post is directly there without the need of manually refreshing.