refactor(challenge-model): merge user methods and generic private methods
This code structure might actually be too suggestive, and it will look weird in main if we decide to do something else for groups. It has been decided not to look so far in advance. This improves readability for the user methods. The helper private methods still think in terms of parentRef. Don't hesitate to revert this if you want implement groups in exactly the same way.