fix(delete-post): fix a use after dispose of state hook
Fixes in two different ways an error that was thrown on post deletion:
`Once you have called dispose() on a ValueNotifier<LoadingState>, it can no longer be used.` for `LoadingIconButton.build`.

1. On the viewmodel `deletePost`, do not await the refresh to enable
faster user feedback (has a side effect of fixing the bug).
2. Check that the widget is still mounted (was not disposed) after
calling the `onClick` function on the `LoadingIconButton` (really fixes
the bug).

Thanks to @yoannLafore for discovering the bug.